+++ astropy/io/fits/tests/test_uint.py
@@ -35,7 +35,7 @@ class TestUintFunctions(FitsTestCase):
             hdu.scale('int{0:d}'.format(bits), '', bzero=2 ** (bits-1))
 
             with ignore_warnings():
-                hdu.writeto(self.temp('tempfile.fits'), clobber=True)
+                hdu.writeto(self.temp('tempfile.fits'), overwrite=True)
 
             with fits.open(self.temp('tempfile.fits'), uint=True) as hdul:
                 assert hdul[hdu_number].data.dtype == self.utype_map[utype]
@@ -89,7 +89,7 @@ class TestUintFunctions(FitsTestCase):
             hdulist = fits.HDUList([hdu0,table])
 
             with ignore_warnings():
-                hdulist.writeto(self.temp('tempfile.fits'), clobber=True)
+                hdulist.writeto(self.temp('tempfile.fits'), overwrite=True)
 
             # Test write of unsigned int
             del hdulist
@@ -103,7 +103,7 @@ class TestUintFunctions(FitsTestCase):
             v = u.view(dtype=[(utype, self.utype_map[utype])])
 
             with ignore_warnings():
-                fits.writeto(self.temp('tempfile2.fits'), v, clobber=True)
+                fits.writeto(self.temp('tempfile2.fits'), v, overwrite=True)
 
             with fits.open(self.temp('tempfile2.fits'), uint=True) as hdulist3:
                 hdudata3 = hdulist3[1].data